Split ListResultBuilder; Change namespace
