+++ DefaultSettings.php
@@ -692,15 +692,15 @@ return array(
 	'smwgResultFormats' => array(
 		'table'      => 'SMW\Query\ResultPrinters\TableResultPrinter',
 		'broadtable' => 'SMW\Query\ResultPrinters\TableResultPrinter',
-		'list'       => 'SMW\ListResultPrinter',
-		'plainlist'  => 'SMW\ListResultPrinter',
-		'ol'         => 'SMW\ListResultPrinter',
-		'ul'         => 'SMW\ListResultPrinter',
+		'list'       => 'SMW\Query\ResultPrinters\ListResultPrinter',
+		'plainlist'  => 'SMW\Query\ResultPrinters\ListResultPrinter',
+		'ol'         => 'SMW\Query\ResultPrinters\ListResultPrinter',
+		'ul'         => 'SMW\Query\ResultPrinters\ListResultPrinter',
 		'category'   => 'SMW\CategoryResultPrinter',
 		'embedded'   => 'SMW\EmbeddedResultPrinter',
-		'template'   => 'SMW\ListResultPrinter',
-		'count'      => 'SMW\ListResultPrinter',
-		'debug'      => 'SMW\ListResultPrinter',
+		'template'   => 'SMW\Query\ResultPrinters\ListResultPrinter',
+		'count'      => 'SMW\Query\ResultPrinters\ListResultPrinter',
+		'debug'      => 'SMW\Query\ResultPrinters\ListResultPrinter',
 		'feed'       => 'SMW\FeedResultPrinter',
 		'csv'        => 'SMW\Query\ResultPrinters\CsvFileExportPrinter',
 		'templatefile' => 'SMW\Query\ResultPrinters\TemplateFileExportPrinter',
