+++ includes/IdeAliases.php
@@ -56,7 +56,7 @@ class SMWEmbeddedResultPrinter extends SMW\EmbeddedResultPrinter {
 class SMWRDFResultPrinter extends SMW\RdfResultPrinter {
 }
 
-class SMWListResultPrinter extends SMW\ListResultPrinter {
+class SMWListResultPrinter extends SMW\Query\ResultPrinters\ListResultPrinter {
 }
 
 interface SMWIResultPrinter extends SMW\QueryResultPrinter {
