--- src/Query/ResultPrinters/ListResultBuilder.php
@@ -1,533 +1,0 @@
-<?php
-
-namespace SMW\Query\ResultPrinters;
-
-use Linker;
-use SMW\ApplicationFactory;
-use SMW\MediaWiki\Renderer\WikitextTemplateRenderer;
-use SMW\Message;
-use SMWDataValue;
-use SMWQueryResult;
-use SMWResultArray;
-
-/**
- * Class ListResultBuilder
- * @package SMWQuery
- */
-class ListResultBuilder {
-
-	private static $defaultConfigurations = [
-		'*' => [
-			'value-open-tag' => '<span class="smw-value">',
-			'value-close-tag' => '</span>',
-			'field-open-tag' => '<span class="smw-field">',
-			'field-close-tag' => '</span>',
-			'field-label-open-tag' => '<span class="smw-field-label">',
-			'field-label-close-tag' => '</span>',
-			'field-label-separator' => ': ',
-			'other-fields-open' => ' (',
-			'other-fields-close' => ')',
-		],
-		'list' => [
-			'row-open-tag' => '<span class="smw-row">',
-			'row-close-tag' => '</span>',
-			'result-open-tag' => '<span class="smw-format list-format">',
-			'result-close-tag' => '</span>',
-		],
-		'ol' => [
-			'row-open-tag' => "<li class=\"smw-row\">",
-			'row-close-tag' => '</li>',
-			'result-open-tag' => '<ol class="smw-format ol-format" start="$START$">',
-			'result-close-tag' => '</ol>',
-		],
-		'ul' => [
-			'row-open-tag' => '<li class="smw-row">',
-			'row-close-tag' => '</li>',
-			'result-open-tag' => '<ul class="smw-format ul-format">',
-			'result-close-tag' => '</ul>',
-		],
-		'plainlist' => [
-			'value-open-tag' => '',
-			'value-close-tag' => '',
-			'field-open-tag' => '',
-			'field-close-tag' => '',
-			'field-label-open-tag' => '',
-			'field-label-close-tag' => '',
-			'row-open-tag' => '',
-			'row-close-tag' => '',
-			'result-open-tag' => '',
-			'result-close-tag' => '',
-		],
-	];
-
-	private $linker = null;
-	private $configuration = [];
-
-	/**
-	 * @var SMWQueryResult
-	 */
-	private $queryResult;
-
-	private $numberOfPages = null;
-	private $templateRenderer;
-
-	/**
-	 * ListResultBuilder constructor.
-	 *
-	 * @param SMWQueryResult $queryResult
-	 * @param Linker $linker
-	 */
-	public function __construct( SMWQueryResult $queryResult, Linker $linker ) {
-		$this->queryResult = $queryResult;
-		$this->linker = $linker;
-	}
-
-	/**
-	 * @param string|string[] $setting
-	 * @param string|null $value
-	 */
-	public function set( $setting, $value = null ) {
-
-		if ( !is_array( $setting ) ) {
-			$setting = [ $setting => $value ];
-		}
-
-		$this->configuration = array_replace( $this->configuration, $setting );
-	}
-
-	/**
-	 * @param string $setting
-	 * @param string $default
-	 *
-	 * @return mixed
-	 */
-	private function get( $setting, $default = '' ) {
-		return isset( $this->configuration[ $setting ] ) ? $this->configuration[ $setting ] : $default;
-	}
-
-	/**
-	 * @return SMWQueryResult
-	 */
-	private function getQueryResult() {
-		return $this->queryResult;
-	}
-
-	/**
-	 * @param $parameter
-	 *
-	 * @return mixed
-	 */
-
-	/**
-	 * @return bool
-	 */
-	private function isSimpleList() {
-		$format = $this->get( 'format' );
-		return $format !== 'ul' && $format !== 'ol' ;
-	}
-
-	/**
-	 * @return string
-	 */
-	public function getResultText() {
-
-		$this->prepareBuilt();
-
-		return
-			$this->getTemplateCall( 'introtemplate' ) .
-			$this->get( 'result-open-tag' ) .
-
-			join( $this->get( 'sep' ), $this->getRowTexts() ) .
-
-			$this->get( 'result-close-tag' ) .
-			$this->getTemplateCall( 'outrotemplate' );
-	}
-
-	/**
-	 * @param string $subject
-	 *
-	 * @return string
-	 */
-	private function insertOffset( $subject ) {
-		return str_replace( '$START$', $this->get( 'offset' ) + 1, $subject );
-	}
-
-	private function prepareBuilt() {
-
-		$format = $this->getEffectiveFormat();
-
-		$this->configuration = array_merge(
-			self::$defaultConfigurations[ '*' ],
-			self::$defaultConfigurations[ $format ],
-			$this->getDefaultsFromI18N(),
-			$this->configuration );
-
-		if ( $this->get( 'template' ) !== '' ) {
-
-			$this->set( [ 'value-open-tag' => '', 'value-close-tag' => '' ] );
-
-		}
-
-		$this->set( 'result-open-tag', $this->insertOffset( $this->get( 'result-open-tag' ) ) );
-	}
-
-	/**
-	 * @return string
-	 */
-	private function getEffectiveFormat() {
-
-		$format = $this->get( 'format' );
-
-		if ( in_array( $format, [ 'ol', 'ul', 'plainlist' ] ) ) {
-			return $format;
-		}
-
-		if ( $this->get( 'template' ) !== '' ) {
-			return 'plainlist';
-		}
-
-		return 'list';
-	}
-
-	/**
-	 * @return string[]
-	 */
-	private function getDefaultsFromI18N() {
-		return [
-			'field-label-separator' => Message::get( 'smw-format-list-field-label-separator' ),
-			'other-fields-open' => Message::get( 'smw-format-list-other-fields-open' ),
-			'other-fields-close' => Message::get( 'smw-format-list-other-fields-close' ),
-		];
-	}
-
-	/**
-	 * @return string[]
-	 */
-	private function getRowTexts() {
-
-		$rowTextFunctionName = $this->get( 'template' ) === '' ? 'getRowText' : 'getRowTextFromTemplate';
-
-		$queryResult = $this->getQueryResult();
-		$queryResult->reset();
-
-		$rowTexts = [];
-		$num = $queryResult->getQuery()->getOffset();
-
-		while ( ( $row = $queryResult->getNext() ) !== false ) {
-
-			$rowTexts[] =
-				$this->get( 'row-open-tag' ) .
-				call_user_func( [ $this, $rowTextFunctionName ], $row, $num ) .
-				$this->get( 'row-close-tag' );
-
-			$num++;
-		}
-
-		return $rowTexts;
-	}
-
-	/**
-	 * @param SMWResultArray[] $fields
-	 *
-	 * @return string
-	 */
-	private function getRowText( array $fields ) {
-
-		$fieldTexts = $this->getFieldTexts( $fields );
-
-		$firstFieldText = array_shift( $fieldTexts );
-
-		if ( $firstFieldText === null ) {
-			return '';
-		}
-
-		if ( count( $fieldTexts ) > 0 ) {
-
-			$otherFieldsText =
-				$this->get( 'other-fields-open' ) .
-				join( $this->get( 'propsep' ), $fieldTexts ) .
-				$this->get( 'other-fields-close' );
-
-		} else {
-			$otherFieldsText = '';
-		}
-
-		return
-			$firstFieldText .
-			$otherFieldsText;
-	}
-
-	/**
-	 * Returns text for one result row, formatted as a template call.
-	 *
-	 * @param SMWResultArray[] $fields
-	 * @param int $rownum
-	 *
-	 * @return string
-	 */
-	private function getRowTextFromTemplate( array $fields, $rownum = 0 ) {
-
-		$templateRenderer = $this->getTemplateRenderer();
-
-		foreach ( $fields as $column => $field ) {
-
-			$fieldLabel = $this->getFieldLabelForTemplate( $field, $column );
-			$fieldText = $this->getValuesText( $field, $column );
-
-			$templateRenderer->addField( $fieldLabel, $fieldText );
-		}
-
-		/** @deprecated since SMW 3.0 */
-		$templateRenderer->addField( '#', $rownum );
-
-		$templateRenderer->addField( '#rownumber', $rownum + 1 );
-		$templateRenderer->packFieldsForTemplate( $this->get( 'template' ) );
-
-		return $templateRenderer->render();
-
-	}
-
-	/**
-	 * @param string[] $fields
-	 *
-	 * @return array
-	 */
-	private function getFieldTexts( array $fields ) {
-
-		$columnNumber = 0;
-		$fieldTexts = [];
-
-		foreach ( $fields as $field ) {
-
-			$valuesText = $this->getValuesText( $field, $columnNumber );
-
-			if ( $valuesText !== '' ) {
-				$fieldTexts[] =
-					$this->get( 'field-open-tag' ) .
-					$this->getFieldLabel( $field ) .
-					$valuesText .
-					$this->get( 'field-close-tag' );
-			}
-
-			$columnNumber++;
-		}
-
-		return $fieldTexts;
-	}
-
-	/**
-	 * @param SMWResultArray $field
-	 *
-	 * @return string
-	 */
-	private function getFieldLabel( SMWResultArray $field ) {
-
-		$showHeaders = $this->get( 'show-headers' );
-
-		if ( $showHeaders === SMW_HEADERS_HIDE || $field->getPrintRequest()->getLabel() === '' ) {
-			return '';
-		}
-
-		$linker = $showHeaders === SMW_HEADERS_PLAIN ? null : $this->linker;
-
-		return
-			$this->get( 'field-label-open-tag' ) .
-			$field->getPrintRequest()->getText( SMW_OUTPUT_WIKI, $linker ) .
-			$this->get( 'field-label-close-tag' ) .
-			$this->get( 'field-label-separator' );
-
-	}
-
-	/**
-	 * @param SMWResultArray $field
-	 * @param int $column
-	 *
-	 * @return string
-	 */
-	private function getFieldLabelForTemplate( SMWResultArray $field, $column ) {
-
-		if ( $this->get( 'named args' ) === false ) {
-			return intval( $column + 1 );
-		}
-
-		$label = $field->getPrintRequest()->getLabel();
-
-		if ( $label === '' ) {
-			return intval( $column + 1 );
-		}
-
-		return $label;
-	}
-
-	/**
-	 * @param SMWResultArray $field
-	 * @param int $column
-	 *
-	 * @return string
-	 */
-	private function getValuesText( SMWResultArray $field, $column = 0 ) {
-
-		$valueTexts = $this->getValueTexts( $field, $column );
-
-		return join( $this->get( 'valuesep' ), $valueTexts );
-
-	}
-
-	/**
-	 * @param SMWResultArray $field
-	 * @param int $column
-	 *
-	 * @return string[]
-	 */
-	private function getValueTexts( SMWResultArray $field, $column ) {
-
-		$valueTexts = [];
-
-		$field->reset();
-
-		while ( ( $dataValue = $field->getNextDataValue() ) !== false ) {
-
-			$valueTexts[] =
-				$this->get( 'value-open-tag' ) .
-				$this->getValueText( $dataValue, $column ) .
-				$this->get( 'value-close-tag' );
-		}
-
-		return $valueTexts;
-	}
-
-	/**
-	 * @param SMWDataValue $value
-	 * @param int $column
-	 *
-	 * @return string
-	 */
-	private function getValueText( SMWDataValue $value, $column = 0 ) {
-
-		$text = $value->getShortText( SMW_OUTPUT_WIKI, $this->getLinkerForColumn( $column ) );
-
-		return $this->sanitizeValueText( $text ) ;
-	}
-
-
-	/**
-	 * @param $text
-	 *
-	 * @return string
-	 */
-	private function sanitizeValueText( $text ) {
-
-		if ( $this->isSimpleList() ) {
-			return $text;
-		}
-
-		return \Sanitizer::removeHTMLtags( $text, null, [], [], [ 'table', 'tr', 'th', 'td', 'dl', 'dd', 'ul', 'li', 'ol' ] );
-	}
-
-	/**
-	 * Depending on current linking settings, returns a linker object
-	 * for making hyperlinks or NULL if no links should be created.
-	 *
-	 * @param int $columnNumber Column number
-	 *
-	 * @return \Linker|null
-	 */
-	private function getLinkerForColumn( $columnNumber ) {
-
-		if ( ( $columnNumber === 0 && $this->get( 'link-first' ) ) ||
-			( $columnNumber > 0 && $this->get( 'link-others' ) ) ) {
-			return $this->linker;
-		}
-
-		return null;
-	}
-
-	/**
-	 * @param string $param
-	 *
-	 * @return string
-	 */
-	private function getTemplateCall( $param ) {
-
-		$templatename = $this->get( $param );
-
-		if ( $templatename === '' ) {
-			return '';
-		}
-
-		$templateRenderer = $this->getTemplateRenderer();
-		$templateRenderer->packFieldsForTemplate( $templatename );
-
-		return $templateRenderer->render();
-
-	}
-
-	/**
-	 * @return WikitextTemplateRenderer
-	 */
-	private function getTemplateRenderer() {
-
-		if ( $this->templateRenderer === null ) {
-			$this->templateRenderer = ApplicationFactory::getInstance()->newMwCollaboratorFactory()->newWikitextTemplateRenderer();
-			$this->addCommonTemplateFields( $this->templateRenderer );
-		}
-
-		return clone( $this->templateRenderer );
-	}
-
-	/**
-	 * @param WikitextTemplateRenderer $templateRenderer
-	 */
-	private function addCommonTemplateFields( WikitextTemplateRenderer $templateRenderer ) {
-
-		$userParam = $this->get( 'userparam' );
-
-		if ( $userParam !== '' ) {
-
-			$templateRenderer->addField(
-				'#userparam',
-				$userParam
-			);
-		}
-
-		$templateRenderer->addField(
-			'#querycondition',
-			$this->getQueryResult()->getQuery()->getQueryString()
-		);
-
-		$templateRenderer->addField(
-			'#querylimit',
-			$this->getQueryResult()->getQuery()->getLimit()
-		);
-
-		$templateRenderer->addField(
-			'#resultoffset',
-			$this->getQueryResult()->getQuery()->getOffset()
-		);
-
-		$templateRenderer->addField(
-			'#rowcount',
-			$this->getRowCount()
-		//$this->getQueryResult()->getCount()  // FIXME: Re-activate if another query takes too long.
-		);
-	}
-
-	/**
-	 * @return int
-	 */
-	private function getRowCount() {
-
-		if ( $this->numberOfPages === null ) {
-
-			$countQuery = \SMWQueryProcessor::createQuery( $this->getQueryResult()->getQueryString(), \SMWQueryProcessor::getProcessedParams( [] ) );
-			$countQuery->querymode = \SMWQuery::MODE_COUNT;
-
-			$queryResult = $this->getQueryResult()->getStore()->getQueryResult( $countQuery );
-			$this->numberOfPages = $queryResult instanceof \SMWQueryResult ? $queryResult->getCountValue() : $queryResult;
-
-		}
-
-		return $this->numberOfPages;
-	}
-
-}
